[BUGFIX] move delayed job from orderPage to OrderCollectionAdminCard
